// apps/frontend/packages/api/base/codegen/types/Remark.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { RemarkStatus2 } from "./RemarkStatus2";

export interface Remark {
    /**
     * @type integer
    */
    id: number;
    /**
     * @type string
    */
    humanId: string;
    /**
     * @type integer
    */
    correctionId: number;
    /**
     * @description null — замечание от ДТОиР
     * @type integer,null
    */
    cfoId?: number | null;
    /**
     * @type integer
    */
    authorId: number;
    /**
     * @type string, date-time
    */
    createdAt: string;
    /**
     * @type integer,null
    */
    relatedSlotId?: number | null;
    /**
     * @type integer,null
    */
    fileVersionId?: number | null;
    /**
     * @type string
    */
    sheetName: string;
    /**
     * @type string
    */
    rowRef: string;
    /**
     * @type string
    */
    cellRef: string;
    /**
     * @type string
    */
    description: string;
    /**
     * @type string
    */
    requiredAction: string;
    /**
     * @type string
    */
    status: RemarkStatus2;
    /**
     * @type integer,null
    */
    closedById?: number | null;
    /**
     * @type string,null, date-time
    */
    closedAt?: string | null;
    /**
     * @description Код ЦФО либо «ДТОиР»
     * @type string
    */
    issuerLabel: string;
}